+++ src/types/AutoRouterOptions.ts
@@ -1,0 +1,8 @@
+import { RequestHandler } from './RequestHandler'
+import { ResponseHandler } from './ResponseHandler'
+import { RouterOptions } from './RouterOptions'
+
+export type AutoRouterOptions = {
+  missing?: RequestHandler
+  format?: ResponseHandler
+} & RouterOptions
