+++ src/types/CustomRoutes.ts
@@ -1,0 +1,5 @@
+import { Route } from './Route'
+
+export type CustomRoutes<R = Route> = {
+  [key: string]: R
+}
