+++ src/types/ErrorHandler.ts
@@ -1,0 +1,4 @@
+import { IRequest } from './IRequest'
+
+export type ErrorHandler<ErrorType = Error, RequestType = IRequest, Args extends any[] = any[]> =
+    (response: ErrorType, request: RequestType, ...args: Args) => any
