+++ src/types/GenericTraps.ts
@@ -1,0 +1,1 @@
+export type GenericTraps = Record<string, any>
