+++ src/types/HasContent.ts
@@ -1,0 +1,5 @@
+import { IRequestStrict } from './IRequestStrict'
+
+export type HasContent<ContentType> = {
+  content: ContentType
+} & IRequestStrict
