+++ src/types/IRequest.ts
@@ -1,0 +1,4 @@
+import { GenericTraps } from './GenericTraps'
+import { IRequestStrict } from './IRequestStrict'
+
+export type IRequest = IRequestStrict & GenericTraps
