+++ src/types/IRequestStrict.ts
@@ -1,0 +1,12 @@
+export type IRequestStrict = {
+  method: string
+  url: string
+  route: string
+  params: {
+    [key: string]: string
+  }
+  query: {
+    [key: string]: string | string[] | undefined
+  }
+  proxy?: any
+} & Request
