+++ src/types/IttyRouterOptions.ts
@@ -1,0 +1,6 @@
+import { RouteEntry } from './RouteEntry'
+
+export type IttyRouterOptions = {
+  base?: string
+  routes?: RouteEntry[]
+} & Record<string, any>
