+++ src/types/RequestHandler.ts
@@ -1,0 +1,4 @@
+import { IRequest } from './IRequest'
+
+export type RequestHandler<R = IRequest, Args extends Array<any> = any[]> =
+  (request: R, ...args: Args) => any
