+++ src/types/RequestLike.ts
@@ -1,0 +1,6 @@
+import { GenericTraps } from './GenericTraps'
+
+export type RequestLike = {
+  method: string
+  url: string
+} & GenericTraps
