+++ src/types/ResponseFormatter.ts
@@ -1,0 +1,3 @@
+export interface ResponseFormatter {
+  (body?: any, options?: ResponseInit): Response
+}
