+++ src/types/ResponseHandler.ts
@@ -1,0 +1,4 @@
+import { IRequest } from './IRequest'
+
+export type ResponseHandler<ResponseType = Response, RequestType = IRequest, Args extends any[] = any[]> =
+    (response: ResponseType & any, request: RequestType & any, ...args: Args) => any
