+++ src/types/RouteEntry.ts
@@ -1,0 +1,8 @@
+import { RequestHandler } from './RequestHandler'
+
+export type RouteEntry = [
+  httpMethod: string,
+  match: RegExp,
+  handlers: RequestHandler[],
+  path?: string,
+]
