+++ src/types/RouterType.ts
@@ -1,0 +1,11 @@
+import { ErrorHandler } from './ErrorHandler'
+import { IRequest } from './IRequest'
+import { IttyRouterType } from './IttyRouterType'
+import { RequestHandler } from './RequestHandler'
+import { ResponseHandler } from './ResponseHandler'
+
+export type RouterType<R = IRequest, Args extends any[] = any[]> = {
+  before?: RequestHandler<any>[]
+  catch?: ErrorHandler
+  finally?: ResponseHandler[]
+} & IttyRouterType<R, Args>
