+++ vitest.config.ts
@@ -10,7 +10,8 @@ export default defineConfig({
         '**/example/**',
         '**/lib/**',
         'src/index.ts',
-        'src/types.d.ts',
+        'src/types',
+        'src/types.ts',
         'src/websocket.ts',
       ],
     },
